+++ src/app/providers/PrivyProviderA.tsx
@@ -1,12 +1,10 @@
 "use client";
 
-import { PrivyClientConfig, PrivyProvider } from "@privy-io/react-auth"
-//import { fuse, sepolia } from "viem/chains";
-import { sepolia, polygonAmoy, polygon, fuse } from "viem/chains"
-import { WagmiProvider, createConfig } from '@privy-io/wagmi';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-// import { Address, Chain, Hash, Transport, http } from "viem"
-import {http} from 'wagmi';
+import { PrivyClientConfig, PrivyProvider } from "@privy-io/react-auth";
+import { sepolia, polygonAmoy, polygon, fuse } from "viem/chains";
+import { WagmiProvider, createConfig } from "@privy-io/wagmi";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { http } from "wagmi";
 
 const handleLogin = (user: any) => {
     console.log(`User ${user.id} logged in!`);
@@ -23,24 +21,26 @@ const queryClient = new QueryClient();
 
 const privyConfig: PrivyClientConfig = {
     embeddedWallets: {
-        createOnLogin: 'users-without-wallets',
+        createOnLogin: "users-without-wallets",
         noPromptOnSignature: false,
     },
-    loginMethods: ['wallet', 'email', 'google'],
+    loginMethods: ["wallet", "email", "google"],
     appearance: {
         showWalletLoginFirst: true,
         theme: "light",
         accentColor: "#676FFF",
         logo: "https://jiffyscan-frontend.vercel.app/images/Frame%2021.svg",
     },
-    supportedChains: [sepolia, polygonAmoy, polygon, fuse
+    supportedChains: [
+        sepolia,
+        polygonAmoy,
+        polygon,
+        fuse,
         // Add any other supported chains here
-      ],
+    ],
 };
 
-export const PrivyProviderA = ({
-    children
-}: { children: React.ReactNode }) => {
+export const PrivyProviderA = ({ children }: { children: React.ReactNode }) => {
     return (
         <PrivyProvider
             // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -56,4 +56,4 @@ export const PrivyProviderA = ({
             </QueryClientProvider>
         </PrivyProvider>
     );
-}
+};
